books/book_shop/views.py
from rest_framework.viewsets import ModelViewSet

from book_shop.models import Author, Book
from book_shop.serializers import AuthorSerializer, BookSerializer


class AuthorViewSet(ModelViewSet):
    queryset = Author.objects.all()
    serializer_class = AuthorSerializer


class BookViewSet(ModelViewSet):
    queryset = Book.objects.all()
    serializer_class = BookSerializer

